Introduced OperatorDict, cleaned-up code significantly
